# docs/swagger.yaml
basePath: /accounts
definitions:
...
  controllers.ListingResp:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Statements'
        type: array
      msg:
        type: string
    type: object
...
  models.Statements:
    properties:
      closed_pnl:
        type: number
      created_time:
        type: string
      exchange:
        type: string
      position_id:
        type: integer
      profit_usd:
        type: number
      quote_amount:
        type: number
      side:
        type: string
      size:
        type: number
      symbol:
        type: string
      updated_time:
        type: string
      user_email:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is BidBox Accounts Service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BidBox Accounts Service API
  version: "1.0"
paths:
...
  /listing/coinwise:
    get:
      consumes:
      - application/json
      description: Coinwise Statements
      parameters:
      - description: service
        in: query
        name: service
        required: true
        type: string
      - description: email
        in: query
        name: email
        required: true
        type: string
      - description: time
        in: query
        name: time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ListingResp'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Coinwise Statements
      tags:
      - listing
...
swagger: "2.0"